Rename models to model
