+++ model/word2vec/huffman/node.go
