+++ model/word2vec/model.go
