+++ model/word2vec/optimizer.go
