+++ model/word2vec/submodel/cbow.go
@@ -12,17 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package model
+package submodel
 
 import (
-	"github.com/ynqa/word-embedding/models"
-	"github.com/ynqa/word-embedding/models/word2vec"
+	"github.com/ynqa/word-embedding/model"
+	"github.com/ynqa/word-embedding/model/word2vec"
 	"github.com/ynqa/word-embedding/utils/vector"
 )
 
 // CBOW is a piece of word2vec model.
 type CBOW struct {
-	models.Common
+	model.Common
 }
 
 // Train updates words' vector using CBOW.
