+++ model/word2vec/word.go
