+++ model/word2vec/word_test.go
